requirements.txt
flask==3.0.0
requests==2.31.0
python-consul==1.1.0